\name{SeedCalc-package}
\alias{SeedCalc-package}
\alias{SeedCalc}
\docType{package}
\title{
Seed germination and seedling growth indexes
}
\description{
Functions to calculate seed germination and seedling emergence and growth indexes.

The main indexes for germination and seedling emergence, considering the time for seed germinate are: T10, T50 (Farooq et al., 2005), T90 e MGT (Labouriau, 1983); considering the germination speed are: Germination Speed Index (Maguire, 1962), Mean Germination Rate (Labouriau, 1983); considering the homogeneity of germination are: Coefficient of Variation of the Germination Time (Carvalho et al., 2005) and Variance of Germination (Labouriau, 1983); Uncertainty (Labouriau & Valadares, 1976)  and Synchrony (Primack, 1980).

The main seedling indexes are Growth, Uniformity and Vigour (Sako, 2001; Medeiros & Pereira, 2018).

}
\details{
\tabular{ll}{
Package: \tab SeedCalc\cr
Type: \tab Package\cr
Version: \tab 1.0.0\cr
Date: \tab 2018-10-31\cr
License: \tab GPL (>= 2)\cr
}
}

\author{
Laercio Junio da Silva \email{laerciojdsilva@gmail.com.br}

Andre Dantas de Medeiros \email{medeiros.seeds@gmail.com }

Ariadne Morbeck Santos Oliveira  \email{ariadneoliveira86@gmail.com}

Maintainer: Laercio Junio da Silva \email{laerciojdsilva@gmail.com}
}

\references{
CARVALHO, M. P., SANTANA, D. G., RANAL, M. A. (2005) <doi:10.1590/S0100-84042005000300018>

FAROOQ, M., BASRA, S. M. A., AHMAD, N., HAFEEZ, K. (2005) <doi:10.1111/j.1744-7909.2005.00031.x>

LABOURIAU L. G., VALADARES, M.B. (1976, ISSN:0001-3765)

LABOURIAU, L. G. (1983) Uma nova linha de pesquisa na fisiologia da germinação das sementes. Anais do XXXIV Congresso Nacional de Botânica. SBB, Porto Alegre, 11–50.

MAGUIRE, J. D. (1962) <doi:10.2135/cropsci1962.0011183X000200020033x>

MEDEIROS, A. D. DE, PEREIRA, M. D. (2018) <doi:10.1590/1983-40632018v4852340>

PRIMACK, R.B. (1980) <doi:10.2307/2259460>

SAKO, Y., MCDONALD, M. B., FUJIMURA, K., EVANS, A. F., BENNETT, M. A. A system for automated seed vigour assessment. Seed Science and Technology, v. 29, n. 3, p. 625–636, 2001.
}

\keyword{ package }
